🚧 Move App.js to containers
